verc/requirements.txt
asn1crypto==0.24.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2018.1.18
cffi==1.14.5
chardet==3.0.4
checksumdir==1.2.0
cryptography==3.4.7
decorator==4.4.2
diskcache==5.1.0
Django==2.2
django-json-rpc==0.7.2
ecdsa==0.16.1
elasticsearch==6.8.1
elasticsearch-dsl==6.4.0
Fabric3==1.14.post1
hashfs==0.7.2
html5lib==1.1
idna==2.6
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
keyring==10.6.0
keyrings.alt==3.0
lxml==4.6.2
memcache==0.4.0
mongoengine==0.22.1
mongokit==0.9.1.1
nodeenv==1.5.0
paramiko==2.7.2
parso==0.7.1
pathlib2==2.3.5
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prompt-toolkit==3.0.10
psycopg2==2.7.4
ptyprocess==0.7.0
pycparser==2.20
pycrypto==2.6.1
Pygments==2.7.4
pygobject==3.26.1
pymongo==3.11.2
PyNaCl==1.4.0
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-magic==0.4.16
python-memcached==1.59
pytz==2020.5
pyxdg==0.25
requests==2.18.4
SecretStorage==2.3.1
semantic-version==2.8.5
setuptools-rust==0.11.6
six==1.11.0
soupsieve==2.1
sqlparse==0.4.1
ssh-import-id==5.7
toml==0.10.2
traitlets==4.3.3
translate-toolkit==3.3.0
unattended-upgrades==0.1
urllib3==1.22
uWSGI==2.0.19.1
wcwidth==0.2.5
webencodings==0.5.1
